move schedule model into src/db
